Move Network Commissioning driver source includes into app projects
Network commissioning source is unconditionally included in the Linux
platform.  But this only makes sense for commissionable apps that
actually use this code.

This commit moves these source includes to the apps and also fixes a
build error that is arising from blindly assigning size_t to a uint8_t.

+++ src/platform/Linux/NetworkCommissioningDriver.h
@@ -180,8 +180,8 @@ public:
 
         // Public, but cannot be accessed via NetworkIterator interface.
         uint8_t interfaceName[kMaxNetworkIDLen];
-        uint8_t interfaceNameLen = 0;
-        bool exhausted           = false;
+        size_t interfaceNameLen = 0;
+        bool exhausted          = false;
     };
 
     uint8_t GetMaxNetworks() override { return 1; };
